+++ backend/src/main/java/com/zzang/chongdae/comment/repository/entity/CommentEntity.java
@@ -1,7 +1,7 @@
-package com.zzang.chongdae.comment.repository;
+package com.zzang.chongdae.comment.repository.entity;
 
-import com.zzang.chongdae.common.repository.BaseTimeEntity;
-import com.zzang.chongdae.member.repository.Member;
+import com.zzang.chongdae.common.repository.entity.BaseTimeEntity;
+import com.zzang.chongdae.member.repository.entity.MemberEntity;
 import com.zzang.chongdae.offering.repository.entity.OfferingEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -9,6 +9,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -19,15 +20,16 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @EqualsAndHashCode(of = "id", callSuper = false)
+@Table(name = "comment")
 @Entity
-public class Comment extends BaseTimeEntity {
+public class CommentEntity extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne
-    private Member member;
+    private MemberEntity member;
 
     @ManyToOne
     private OfferingEntity offering;
